requirements.txt
django-model-utils==1.3.1
jsonfield==0.9.15
Mezzanine==1.4.6
